Split `tta` into a separate file
